apps/ui: one folder per route, `layout-` and `route-` prefix convention
Every route lives in its own folder now. Layouts are separate from the
pages they wrap, prefixed `layout-*`, and each addressable page is in a
`route-*` folder:

  router/
    layout-root/              ← rootRoute + RouterContext
    layout-dashboard/         ← SidebarLayout shell
    layout-onboarding/        ← OnboardingLayout shell + shared CSS
    route-index/              ← "/" redirect
    route-dashboard/          ← "/dashboard"
    route-session-detail/     ← "/sessions/$sessionId"
    route-onboarding-home/    ← "/onboarding"
    route-onboarding-create/  ← "/onboarding/create"
    router.tsx                ← tree composition via addChildren

Each folder's index.tsx exports just its own route (and component);
router.tsx owns the wiring so layouts don't need to import their
children. Shared onboarding page CSS moves into layout-onboarding/
since the chrome belongs to the layout.
