+++ apps/ui/src/router/router.tsx
@@ -1,14 +1,18 @@
 import { createRouter } from '@tanstack/react-router';
-import { dashboardLayoutWithChildren } from './dashboard';
-import { onboardingRoute } from './onboarding';
-import { rootRoute } from './root';
-import { indexRoute } from './index';
-import type { RouterContext } from './root';
+import { dashboardLayoutRoute } from './layout-dashboard';
+import { onboardingLayoutRoute } from './layout-onboarding';
+import { rootRoute } from './layout-root';
+import { dashboardRoute } from './route-dashboard';
+import { indexRoute } from './route-index';
+import { onboardingCreateRoute } from './route-onboarding-create';
+import { onboardingHomeRoute } from './route-onboarding-home';
+import { sessionDetailRoute } from './route-session-detail';
+import type { RouterContext } from './layout-root';
 
 const routeTree = rootRoute.addChildren( [
 	indexRoute,
-	dashboardLayoutWithChildren,
-	onboardingRoute,
+	dashboardLayoutRoute.addChildren( [ dashboardRoute, sessionDetailRoute ] ),
+	onboardingLayoutRoute.addChildren( [ onboardingHomeRoute, onboardingCreateRoute ] ),
 ] );
 
 export function createAppRouter( context: RouterContext ) {
